requirements.txt
accelerate==1.10.1
asgiref==3.9.1
babel==2.17.0
breadability==0.1.20
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
courlan==1.3.2
dateparser==1.2.2
Django==5.2.6
djangorestframework==3.16.1
docopt==0.6.2
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
htmldate==1.9.3
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
packaging==25.0
protobuf==6.32.1
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycountry==24.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
six==1.17.0
sqlparse==0.5.3
sumy==0.11.0
sympy==1.14.0
tld==0.13.1
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.56.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
